fix(api): eagerly open media SQLite at boot + wire setMediaDb into setupTestContext (Copilot review)
Two adjustments after Copilot's review of #2791:

1. apps/pops-api/src/index.ts now calls getMediaDrizzle() at boot
   alongside getCoreDrizzle(), so the media pillar's per-pillar
   migrations land before the first request hits the API. The PR 2
   docstring previously claimed this happened — it now does.

2. apps/pops-api/src/shared/test-utils.ts setupTestContext() now calls
   setMediaDb({ db: drizzle(db), raw: db }) alongside setCoreDb so
   tests route both pillar handles at the same in-memory fixture, and
   teardown clears setMediaDb(null). Matches the docstring on setMediaDb.

Verification:

- pnpm --filter @pops/api typecheck — clean
- pnpm --filter @pops/api test src/db/media-sqlite-path.test.ts
  src/db/per-pillar-migrations.test.ts — 11/11 cases pass
- pnpm format:check — clean
